+++ example/lib/main.dart
@@ -34,12 +34,6 @@ class _MyWidget extends StatelessWidget {
     return Scaffold(
       body: Row(
         children: [
-          switch (const AsyncSnapshot<int>.nothing()) {
-            AsyncSnapshot(:final data?) => Center(child: Text(data.toString())),
-            AsyncSnapshot(:final error?) =>
-              Center(child: Text(error.toString())),
-            _ => const CircularProgressIndicator(),
-          },
           Column(
             children: [
               // AsyncBuilder(
@@ -55,58 +49,68 @@ class _MyWidget extends StatelessWidget {
               //   },
               // ),
 
-              AsyncButtonBuilder(
+              AsyncButton(
                 onPressed: onError,
                 child: const FlutterLogo(size: 120),
                 builder: (context, state, child) {
                   return InkWell(
-                    onTap: state.press,
+                    // onTap: state.press,
                     child: child,
                   );
                 },
               ),
               ElevatedButton(
                 onPressed: onError,
+                onLongPress: onSuccess,
                 child: const Text('ElevatedButton'),
               ).asAsync(),
               ElevatedButton.icon(
                 onPressed: onError,
+                onLongPress: onSuccess,
                 label: const Text('ElevatedButton.icon'),
                 icon: const Icon(Icons.add),
               ).asAsync(),
               FilledButton(
                 onPressed: onError,
+                onLongPress: onSuccess,
                 child: const Text('FilledButton'),
               ).asAsync(),
               FilledButton.icon(
                 onPressed: onError,
+                onLongPress: onSuccess,
                 label: const Text('FilledButton.icon'),
                 icon: const Icon(Icons.add),
               ).asAsync(),
               FilledButton.tonal(
                 onPressed: onError,
+                onLongPress: onSuccess,
                 child: const Text('FilledButton.tonal'),
               ).asAsync(),
               FilledButton.tonalIcon(
                 onPressed: onError,
+                onLongPress: onSuccess,
                 label: const Text('FilledButton.tonalIcon'),
                 icon: const Icon(Icons.add),
               ).asAsync(),
               OutlinedButton(
                 onPressed: onError,
+                onLongPress: onSuccess,
                 child: const Text('OutlinedButton'),
               ).asAsync(),
               OutlinedButton.icon(
                 onPressed: onError,
+                onLongPress: onSuccess,
                 label: const Text('OutlinedButton.icon'),
                 icon: const Icon(Icons.add),
               ).asAsync(),
               TextButton(
                 onPressed: onError,
+                onLongPress: onSuccess,
                 child: const Text('TextButton'),
               ).asAsync(),
               TextButton.icon(
                 onPressed: onError,
+                onLongPress: onSuccess,
                 label: const Text('TextButton.icon'),
                 icon: const Icon(Icons.add),
               ).asAsync(),
@@ -115,19 +119,24 @@ class _MyWidget extends StatelessWidget {
           Column(
             children: [
               IconButton(
+                key: AsyncKey('value_icon_button'),
                 onPressed: onError,
+                onLongPress: onSuccess,
                 icon: const Icon(Icons.add),
               ).asAsync(),
               IconButton.filled(
                 onPressed: onSuccess,
+                onLongPress: onError,
                 icon: const Icon(Icons.add),
               ).asAsync(successIcon: const Icon(Icons.check)),
               IconButton.filledTonal(
                 onPressed: onError,
+                onLongPress: onSuccess,
                 icon: const Icon(Icons.add),
               ).asAsync(),
               IconButton.outlined(
                 onPressed: onSuccess,
+                onLongPress: onError,
                 icon: const Icon(Icons.add),
               ).asAsync(),
             ],
@@ -180,7 +189,9 @@ class _MyWidget extends StatelessWidget {
               label: const Text('extended'),
             ).asAsync(),
             FloatingActionButton.large(
-              onPressed: onSuccess,
+              onPressed: () async {
+                await AsyncKey('value_icon_button').currentButton?.press();
+              },
               child: const Icon(Icons.add),
             ).asAsync(),
           ].map((e) => Padding(padding: fabPadding, child: e)).toList(),