+++ lib/src/utils/async_state.dart
@@ -1,6 +1,8 @@
 import 'package:async_notifier/async_notifier.dart';
 import 'package:flutter/material.dart';
 
+import '../widgets/async/async.dart';
+
 /// A [State] with an [AsyncNotifier].
 abstract class AsyncState<T extends StatefulWidget, Data> extends State<T> {
   /// The [AsyncNotifier] of this state.
@@ -14,7 +16,10 @@ abstract class AsyncState<T extends StatefulWidget, Data> extends State<T> {
   AsyncSnapshot<Data> get snapshot => async.snapshot;
 
   /// Called whenever [AsyncSnapshot] changes.
-  void onSnapshot(AsyncSnapshot<Data> snapshot) {}
+  @mustCallSuper
+  void onSnapshot(AsyncSnapshot<Data> snapshot) {
+    snapshot.whenOrNull(error: Async.errorLogger);
+  }
 
   @override
   void initState() {